runner/requirements.txt
pyyaml==6.0.1
rich==13.7.0
pandas
tabulate==0.9.0
numpy
kubernetes==29.0.0
requests==2.31.0